Move mood_error_t and mood_t into 'common' lib
