--- homework_2/lib/errors/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(errors INTERFACE)
-target_include_directories(errors INTERFACE include)
