+++ homework_2/lib/mood/CMakeLists.txt
@@ -1,4 +1,4 @@
 set(SOURCES src/mood_determine.c)
 add_library(mood STATIC ${SOURCES})
 target_include_directories(mood PUBLIC include)
-target_link_libraries(mood PUBLIC errors)
+target_link_libraries(mood PUBLIC mood_common)
