+++ homework_2/lib/mood/include/mood/mood_determine.h
@@ -1,10 +1,5 @@
 #pragma once
-#include <errors/errors.h>
-
-typedef enum {
-    POSITIVE,
-    NEUTRAL,
-    NEGATIVE,
-} mood_t;
+#include <mood_common/errors.h>
+#include <mood_common/mood_state.h>
 
 mood_error_t mood_determine(const char *const str, mood_t *out);
