+++ homework_2/lib/mood_common/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(mood_common INTERFACE)
+target_include_directories(mood_common INTERFACE include)
