+++ homework_2/lib/mood_common/include/mood_common/errors.h
