+++ homework_2/lib/mood_common/include/mood_common/mood_state.h
@@ -1,0 +1,7 @@
+#pragma once
+
+typedef enum {
+    POSITIVE,
+    NEUTRAL,
+    NEGATIVE,
+} mood_t;
