+++ homework_2/lib/mood_mp/CMakeLists.txt
@@ -1,4 +1,4 @@
 set(SOURCES src/test_func.c)
 add_library(mood_mp SHARED ${SOURCES})
 target_include_directories(mood_mp PUBLIC include)
-target_link_libraries(mood_mp PUBLIC errors)
+target_link_libraries(mood_mp PUBLIC mood_common)
