+++ homework_2/lib/reader_utils/CMakeLists.txt
@@ -1,4 +1,4 @@
 set(SOURCES src/read_file.c)
 add_library(reader_utils ${SOURCES})
 target_include_directories(reader_utils PUBLIC include)
-target_link_libraries(reader_utils PUBLIC errors)
+target_link_libraries(reader_utils PUBLIC mood_common)
