+++ homework_2/lib/reader_utils/include/reader_utils/read_file.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <stdio.h>
 #include <stdlib.h>
-#include "errors/errors.h"
+#include "mood_common/mood_state.h"
+#include "mood_common/errors.h"
 
 mood_error_t read_file(const char *const filepath, char **out_str);
